Refactor debounce function and add autocomplete functionality to multiple elements
