+++ index.html
@@ -35,6 +35,8 @@
     </section>
     <div class="container">
       <div class="autocomplete"></div>
+      <div class="autocomplete-two"></div>
+      <div class="autocomplete-three"></div>
       <div id="summary"></div>
     </div>
     <!-- <script src="utils.js"></script> -->
